Alpha 0.20.0
Changed the layout. Now has a login and a register functionality. The
old verson is completely put in the folder _oldKomodo (if you forget how
to old stucture worked). Certain files like script.js are already placed
in the js folder of the new layout just a images and styling.